refactor: Move decoder evm calls to their own file.
